fixed idea

+++ src/main/java/g3601_3700/s3695_maximize_alternating_sum_using_swaps/Solution.java
@@ -36,7 +36,7 @@ public class Solution {
             }
             list[i].sort((a, b) -> a - b);
             for (int j = 0; j < list[i].size(); j++) {
-                result += list[i].get(j) * (j < oddCount[r] ? -1 : 1);
+                result += (long) list[i].get(j) * (j < oddCount[r] ? -1 : 1);
             }
         }
         return result;
